+++ native_library/android/FacebookANE/app/src/main/java/com/tuarua/facebookane/KotlinController.kt
@@ -24,7 +24,11 @@ import com.adobe.fre.FREContext
 import com.adobe.fre.FREObject
 import com.tuarua.frekotlin.*
 import android.content.pm.PackageManager
+import android.content.res.Configuration
 import android.util.Base64
+import com.adobe.air.AndroidActivityWrapper
+import com.adobe.air.FreKotlinActivityResultCallback
+import com.adobe.air.FreKotlinStateChangeCallback
 import com.facebook.*
 import com.facebook.login.DefaultAudience
 import com.facebook.login.LoginBehavior
@@ -50,10 +54,10 @@ import com.tuarua.facebookane.extensions.toMap
 import java.util.*
 
 @Suppress("unused", "UNUSED_PARAMETER", "UNCHECKED_CAST", "DEPRECATION")
-class KotlinController : FreKotlinMainController {
-    private var onShareSuccessEventId: String? = null
-    private var onShareCancelEventId: String? = null
-    private var onShareErrorEventId: String? = null
+class KotlinController : FreKotlinMainController, FreKotlinStateChangeCallback, FreKotlinActivityResultCallback {
+    private var onShareSuccessCallbackId: String? = null
+    private var onShareCancelCallbackId: String? = null
+    private var onShareErrorCallbackId: String? = null
     private var callbackManager: CallbackManager? = null
     private lateinit var accessTokenTracker: AccessTokenTracker
     private lateinit var activity: Activity
@@ -68,9 +72,9 @@ class KotlinController : FreKotlinMainController {
     // FacebookSdk
 
     fun init(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 3 } ?: return FreArgException("init")
+        argv.takeIf { argv.size > 3 } ?: return FreArgException()
         val applicationId = String(argv[0]) ?: return null
-        val onCurrentAccessTokenChangedEventId = String(argv[1])
+        val onCurrentAccessTokenChangedCallbackId = String(argv[1])
         val isAdvertiserIDCollectionEnabled = Boolean(argv[2]) == true
         val isAutoLogAppEventsEnabled = Boolean(argv[3]) == true
 
@@ -83,11 +87,11 @@ class KotlinController : FreKotlinMainController {
         FacebookSdk.setAutoLogAppEventsEnabled(isAutoLogAppEventsEnabled)
 
         callbackManager = CallbackManager.Factory.create()
-        if (onCurrentAccessTokenChangedEventId != null) {
+        if (onCurrentAccessTokenChangedCallbackId != null) {
             accessTokenTracker = object : AccessTokenTracker() {
                 override fun onCurrentAccessTokenChanged(oldToken: AccessToken?, newToken: AccessToken?) {
                     dispatchEvent(FacebookEvent.ON_CURRENT_ACCESS_TOKEN_CHANGED,
-                            Gson().toJson(FacebookEvent(onCurrentAccessTokenChangedEventId,
+                            Gson().toJson(FacebookEvent(onCurrentAccessTokenChangedCallbackId,
                                     mapOf("oldToken" to oldToken.toMap(), "newToken" to newToken.toMap())))
                     )
                 }
@@ -102,7 +106,7 @@ class KotlinController : FreKotlinMainController {
     }
 
     fun setIsDebugEnabled(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 0 } ?: return FreArgException("setIsDebugEnabled")
+        argv.takeIf { argv.size > 0 } ?: return FreArgException()
         FacebookSdk.setIsDebugEnabled(Boolean(argv[0]) == true)
         return null
     }
@@ -112,13 +116,13 @@ class KotlinController : FreKotlinMainController {
     }
 
     fun addLoggingBehavior(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 0 } ?: return FreArgException("addLoggingBehavior")
+        argv.takeIf { argv.size > 0 } ?: return FreArgException()
         FacebookSdk.addLoggingBehavior(loggingBehaviorFromInt(Int(argv[0]) ?: 0))
         return null
     }
 
     fun removeLoggingBehavior(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 0 } ?: return FreArgException("removeLoggingBehavior")
+        argv.takeIf { argv.size > 0 } ?: return FreArgException()
         FacebookSdk.removeLoggingBehavior(loggingBehaviorFromInt(Int(argv[0]) ?: 0))
         return null
     }
@@ -129,7 +133,7 @@ class KotlinController : FreKotlinMainController {
     }
 
     fun isLoggingBehaviorEnabled(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 0 } ?: return FreArgException("isLoggingBehaviorEnabled")
+        argv.takeIf { argv.size > 0 } ?: return FreArgException()
         return FacebookSdk.isLoggingBehaviorEnabled(loggingBehaviorFromInt(Int(argv[0])
                 ?: 0)).toFREObject()
     }
@@ -139,7 +143,7 @@ class KotlinController : FreKotlinMainController {
     }
 
     fun setLimitEventAndDataUsage(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 0 } ?: return FreArgException("removeLoggingBehavior")
+        argv.takeIf { argv.size > 0 } ?: return FreArgException()
         FacebookSdk.setLimitEventAndDataUsage(activity, Boolean(argv[0]) == true)
         return null
     }
@@ -169,14 +173,13 @@ class KotlinController : FreKotlinMainController {
     }
 
     fun setIsAdvertiserIDCollectionEnabled(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 0 }
-                ?: return FreArgException("setIsAdvertiserIDCollectionEnabled")
+        argv.takeIf { argv.size > 0 } ?: return FreArgException()
         if (FacebookSdk.isInitialized()) FacebookSdk.setAdvertiserIDCollectionEnabled(Boolean(argv[0]) == true)
         return null
     }
 
     fun setIsAutoLogAppEventsEnabled(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 0 } ?: return FreArgException("setIsAutoLogAppEventsEnabled")
+        argv.takeIf { argv.size > 0 } ?: return FreArgException()
         if (FacebookSdk.isInitialized()) FacebookSdk.setAutoLogAppEventsEnabled(Boolean(argv[0]) == true)
         return null
     }
@@ -184,17 +187,17 @@ class KotlinController : FreKotlinMainController {
     // LoginManager
 
     fun login(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 4 } ?: return FreArgException("removeLoggingBehavior")
+        argv.takeIf { argv.size > 4 } ?: return FreArgException()
         val permissions = List<String>(FREArray(argv[0]))
         val withPublish = Boolean(argv[1]) == true
-        val onSuccessEventId = String(argv[2]) ?: return null
-        val onCancelEventId = String(argv[3]) ?: return null
-        val onErrorEventId = String(argv[4]) ?: return null
+        val onSuccessCallbackId = String(argv[2]) ?: return null
+        val onCancelCallbackId = String(argv[3]) ?: return null
+        val onErrorCallbackId = String(argv[4]) ?: return null
 
         LoginManager.getInstance().registerCallback(callbackManager, object : FacebookCallback<LoginResult> {
             override fun onSuccess(loginResult: LoginResult) {
                 dispatchEvent(FacebookEvent.ON_LOGIN_SUCCESS,
-                        Gson().toJson(FacebookEvent(onSuccessEventId,
+                        Gson().toJson(FacebookEvent(onSuccessCallbackId,
                                 mapOf("accessToken" to loginResult.accessToken.toMap(),
                                         "recentlyGrantedPermissions" to loginResult.recentlyGrantedPermissions,
                                         "recentlyDeniedPermissions" to loginResult.recentlyDeniedPermissions)))
@@ -204,14 +207,14 @@ class KotlinController : FreKotlinMainController {
 
             override fun onCancel() {
                 dispatchEvent(FacebookEvent.ON_LOGIN_CANCEL,
-                        Gson().toJson(FacebookEvent(onCancelEventId))
+                        Gson().toJson(FacebookEvent(onCancelCallbackId))
                 )
                 LoginManager.getInstance().unregisterCallback(callbackManager)
             }
 
             override fun onError(error: FacebookException) {
                 dispatchEvent(FacebookEvent.ON_LOGIN_ERROR,
-                        Gson().toJson(FacebookEvent(onErrorEventId,
+                        Gson().toJson(FacebookEvent(onErrorCallbackId,
                                 mapOf("message" to error.localizedMessage))
                         )
                 )
@@ -232,8 +235,9 @@ class KotlinController : FreKotlinMainController {
         return LoginManager.getInstance().loginBehavior.ordinal.toFREObject()
     }
 
+
     fun setLoginBehavior(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 0 } ?: return FreArgException("setLoginBehavior")
+        argv.takeIf { argv.size > 0 } ?: return FreArgException()
         when (Int(argv[0]) ?: 0) {
             0 -> LoginManager.getInstance().loginBehavior = LoginBehavior.NATIVE_WITH_FALLBACK
             1 -> LoginManager.getInstance().loginBehavior = LoginBehavior.NATIVE_ONLY
@@ -251,7 +255,7 @@ class KotlinController : FreKotlinMainController {
     }
 
     fun setDefaultAudience(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 0 } ?: return FreArgException("setDefaultAudience")
+        argv.takeIf { argv.size > 0 } ?: return FreArgException()
         when (Int(argv[0]) ?: 0) {
             0 -> LoginManager.getInstance().defaultAudience = DefaultAudience.NONE
             1 -> LoginManager.getInstance().defaultAudience = DefaultAudience.ONLY_ME
@@ -269,19 +273,19 @@ class KotlinController : FreKotlinMainController {
     }
 
     fun refreshCurrentAccessTokenAsync(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 1 } ?: return FreArgException("refreshCurrentAccessTokenAsync")
-        val onTokenRefreshedEventId = String(argv[0]) ?: return null
-        val onTokenRefreshFailedEventId = String(argv[1]) ?: return null
+        argv.takeIf { argv.size > 1 } ?: return FreArgException()
+        val onTokenRefreshedCallbackId = String(argv[0]) ?: return null
+        val onTokenRefreshFailedCallbackId = String(argv[1]) ?: return null
         AccessToken.refreshCurrentAccessTokenAsync(object : AccessToken.AccessTokenRefreshCallback {
             override fun OnTokenRefreshed(accessToken: AccessToken) {
                 dispatchEvent(FacebookEvent.ON_TOKEN_REFRESH,
-                        Gson().toJson(FacebookEvent(onTokenRefreshedEventId, accessToken.toMap()))
+                        Gson().toJson(FacebookEvent(onTokenRefreshedCallbackId, accessToken.toMap()))
                 )
             }
 
             override fun OnTokenRefreshFailed(exception: FacebookException) {
                 dispatchEvent(FacebookEvent.ON_TOKEN_REFRESH_FAILED,
-                        Gson().toJson(FacebookEvent(onTokenRefreshFailedEventId,
+                        Gson().toJson(FacebookEvent(onTokenRefreshFailedCallbackId,
                                 mapOf("message" to exception.localizedMessage))
                         )
                 )
@@ -314,20 +318,20 @@ class KotlinController : FreKotlinMainController {
 
     // Share
     fun onShareSuccess(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 0 } ?: return FreArgException("onShareSuccess")
-        onShareSuccessEventId = String(argv[0])
+        argv.takeIf { argv.size > 0 } ?: return FreArgException()
+        onShareSuccessCallbackId = String(argv[0])
         return null
     }
 
     fun onShareCancel(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 0 } ?: return FreArgException("onShareCancel")
-        onShareCancelEventId = String(argv[0])
+        argv.takeIf { argv.size > 0 } ?: return FreArgException()
+        onShareCancelCallbackId = String(argv[0])
         return null
     }
 
     fun onShareError(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 0 } ?: return FreArgException("onShareError")
-        onShareErrorEventId = String(argv[0])
+        argv.takeIf { argv.size > 0 } ?: return FreArgException()
+        onShareErrorCallbackId = String(argv[0])
         return null
     }
 
@@ -341,7 +345,7 @@ class KotlinController : FreKotlinMainController {
     }
 
     fun shareDialog_create(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 1 } ?: return FreArgException("shareDialog_create")
+        argv.takeIf { argv.size > 1 } ?: return FreArgException()
         val id = UUID.randomUUID().toString()
         val dialog = ShareDialog(activity)
         dialog.registerCallback(callbackManager, ResultFacebookCallback(dialog))
@@ -350,7 +354,7 @@ class KotlinController : FreKotlinMainController {
     }
 
     fun shareDialog_show(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 1 } ?: return FreArgException("shareDialog_show")
+        argv.takeIf { argv.size > 1 } ?: return FreArgException()
         val id = String(argv[0]) ?: return null
         val content = createSharingContent(argv[1]) ?: return null
         shareDialogs[id]?.show(content)
@@ -362,7 +366,7 @@ class KotlinController : FreKotlinMainController {
     }
 
     fun messageDialog_create(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 1 } ?: return FreArgException("messageDialog_create")
+        argv.takeIf { argv.size > 1 } ?: return FreArgException()
         val id = UUID.randomUUID().toString()
         val dialog = MessageDialog(activity)
         dialog.registerCallback(callbackManager, ResultFacebookCallback(dialog))
@@ -371,7 +375,7 @@ class KotlinController : FreKotlinMainController {
     }
 
     fun messageDialog_show(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 1 } ?: return FreArgException("messageDialog_show")
+        argv.takeIf { argv.size > 1 } ?: return FreArgException()
         val id = String(argv[0]) ?: return null
         val content = createSharingContent(argv[1]) ?: return null
         messageDialogs[id]?.show(content)
@@ -383,7 +387,7 @@ class KotlinController : FreKotlinMainController {
     }
 
     fun shareAPI_create(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 3 } ?: return FreArgException("shareAPI_create")
+        argv.takeIf { argv.size > 3 } ?: return FreArgException()
         val content = createSharingContent(argv[1]) ?: return null
         val id = UUID.randomUUID().toString()
         val api = ShareApi(content)
@@ -392,27 +396,24 @@ class KotlinController : FreKotlinMainController {
     }
 
     fun shareAPI_share(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 1 } ?: return FreArgException("shareAPI_canShare")
+        argv.takeIf { argv.size > 1 } ?: return FreArgException()
         val id = String(argv[0]) ?: return null
         shareAPIs[id]?.share(ResultFacebookCallback(null))
         return null
     }
 
     fun shareAPI_canShare(ctx: FREContext, argv: FREArgv): FREObject? {
-        argv.takeIf { argv.size > 1 } ?: return FreArgException("shareAPI_canShare")
+        argv.takeIf { argv.size > 1 } ?: return FreArgException()
         val id = String(argv[0]) ?: return null
         return shareAPIs[id]?.canShare()?.toFREObject()
     }
 
-    fun onActivityResult(requestCode: Int, resultCode: Int, intent: Intent) {
-        callbackManager?.onActivityResult(requestCode, resultCode, intent)
-    }
 
     inner class ResultFacebookCallback(private val dialog: FacebookDialogBase<ShareContent<*, *>, Sharer.Result>?) : FacebookCallback<Sharer.Result> {
         override fun onSuccess(result: Sharer.Result?) {
-            if (onShareSuccessEventId != null) {
+            if (onShareSuccessCallbackId != null) {
                 dispatchEvent(FacebookEvent.ON_SHARE_SUCCESS,
-                        Gson().toJson(FacebookEvent(onShareSuccessEventId,
+                        Gson().toJson(FacebookEvent(onShareSuccessCallbackId,
                                 mapOf("postId" to result?.postId)))
                 )
             }
@@ -420,18 +421,18 @@ class KotlinController : FreKotlinMainController {
         }
 
         override fun onCancel() {
-            if (onShareCancelEventId != null) {
+            if (onShareCancelCallbackId != null) {
                 dispatchEvent(FacebookEvent.ON_SHARE_CANCEL,
-                        Gson().toJson(FacebookEvent(onShareCancelEventId))
+                        Gson().toJson(FacebookEvent(onShareCancelCallbackId))
                 )
             }
             dialog?.registerCallback(callbackManager, null)
         }
 
         override fun onError(error: FacebookException?) {
-            if (onShareErrorEventId != null) {
+            if (onShareErrorCallbackId != null) {
                 dispatchEvent(FacebookEvent.ON_SHARE_ERROR,
-                        Gson().toJson(FacebookEvent(onShareErrorEventId,
+                        Gson().toJson(FacebookEvent(onShareErrorCallbackId,
                                 mapOf("message" to error?.localizedMessage))
                         )
                 )
@@ -440,7 +441,18 @@ class KotlinController : FreKotlinMainController {
         }
     }
 
-    override val TAG: String
+    override fun onActivityResult(requestCode: Int, resultCode: Int, intent: Intent?) {
+        callbackManager?.onActivityResult(requestCode, resultCode, intent)
+    }
+
+    override fun onConfigurationChanged(configuration: Configuration?) {
+    }
+
+    override fun onActivityStateChanged(activityState: AndroidActivityWrapper.ActivityState?) {
+
+    }
+
+    override val TAG: String?
         get() = this::class.java.simpleName
     private var _context: FREContext? = null
     override var context: FREContext?