+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### 1.2.0 
+- AND: Updated to jetpack/androidx
+- AND: Updated to FreKotlin 1.9.5
+- iOS: Updated to FreSwift 4.3.0
+- iOS: upgrade to 5.13.1 FBSDK
+- iOS: use Swift Package Manager
+- iOS: MessageDialog is now deprecated
+- Android: upgrade to 5.15.0 FBSDK
+
 ### 1.1.0 
 - add `FacebookSdk.onShareSuccess`
 - add `FacebookSdk.onShareCancel`